Make Migrations work universally

+++ db/migrate/20110902193901_delete_urinetown_events.rb
@@ -1,11 +1,13 @@
 class DeleteUrinetownEvents < ActiveRecord::Migration
   def self.up
-  	g = Show.find 48
-  	e = g.events
-  	f = Feedpost.where(:parent_id => 48).where(:post_type => "create")
+    if Show.exists?(:id => 48)
+  	  g = Show.find 48
+  	  e = g.events
+  	  f = Feedpost.where(:parent_id => 48).where(:post_type => "create")
 
-  	f.destroy_all
-  	e.destroy_all
+  	  f.destroy_all
+  	  e.destroy_all
+	  end
 
   end
